AI Calculation, SAT Math & Verbal error messages
The Academic Index calculation seems to be working with SAT Math &
Verbal scores as input.  If either of the two scores is too low or high,
the composite score now becomes an error message.  The ACT -> SAT
Equivalent part doesn't seem to be factoring into the AI calculation
properly, but I'm not completely sure how that's going to work anyway.