Everything's working, for the most part.
AI is calculated, and when it displays the calculation, you can click a
link to get a popup showing the calculation detail, but I think I'm
going to change the control styles.  The one problem at the moment is
when you try to switch between SAT scores & ACT scores, the calculation
is not completely intuitive.