Select boxes
I changed the SAT/ACT score inputs to select boxes, thinking this makes
it a little easier to check on mobile devices, and it also restricts
values to valid inputs.  Now I like the idea more for the latter, since
it takes a few swipes to get down to the right score, and the numbers to
type aren't long.  Still undecided if this will be the final version of
these controls.