Rounding AI Result to 1 decimal place.
Added a little code to round the result to 1 decimal place, to match the
original spreadsheet.  Also, added the current floor of 183, hard-coded
into calculate.js.  If AI result is NaN (not a number), an error message
pops up.  This occurs if the user uses the back or forward button to
access the page.